+++ tests/integration/test_utils/data_source_config/__init__.py
@@ -1,4 +1,4 @@
-from .base_data_source_config import BatchSetup, DataSourceConfig
+from .base import BatchSetup, DataSourceConfig
 from .pandas_data_frame import PandasDataFrameBatchSetup, PandasDataFrameDatasourceConfig
 
 __all__ = [
